// extensions/solana-agent/src/type.ts
export type Set<T> = React.Dispatch<React.SetStateAction<T>>;
// ...
export interface Question {
  id: string;
  question: string;
  files: string[];
  created_at: string;
}

export interface Chat extends Question {
  answer: string;
}

export interface SavedChat extends Chat {
  saved_at?: string;
}

export interface Conversation {
  id: string;
  model: Model;
  chats: Chat[];
  updated_at: string;
  created_at: string;
  pinned: boolean;
}

export interface Model {
  id: string;
  updated_at: string;
  created_at: string;
  name: string;
  prompt: string;
  option: string;
  temperature: string;
  pinned: boolean;
  vision?: boolean;
}

export type CommandContentSource = "clipboard" | "selectedText" | "browserTab";

export interface Command {
  id: string;
  name: string;
  prompt: string;
  model: string;
  temperature: string;
  contentSource: CommandContentSource;
  isDisplayInput: boolean;
}

type FunctionNoArg = () => void;
type PromiseFunctionNoArg = () => Promise<void>;
type PromiseFunctionWithOneArg<T> = (arg: T) => Promise<void>;
// type PromiseFunctionWithTwoArg<T, V> = (arg_1: T, arg_2: V) => Promise<void>;
type PromiseFunctionWithThreeArg<T, V, W> = (arg_1: T, arg_2: V, arg_3: W) => Promise<void>;
// ...
export interface ChatHook {
  data: Chat[];
  errorMsg: string | null;
  setData: Set<Chat[]>;
  isLoading: boolean;
  isAborted: boolean;
  setLoading: Set<boolean>;
  selectedChatId: string | null;
  setSelectedChatId: Set<string | null>;
  ask: PromiseFunctionWithThreeArg<string, string[], Model>;
  clear: PromiseFunctionNoArg;
  abort: FunctionNoArg;
  streamData: Chat | undefined;
}